Add MsgHook plugin (#1)
* Add MsgHook plugin

Adds the MsgHook plugin to the plugins folder. Also adds a bunch of Node
stuff to allow TypeScript support and Prettier formatting.

* Update plugin description

* Add window.MsgHook.hasCommand method

Adds a method to check if a message starts with a given string and to
return the message without that string. This makes it easier to make
commands that modify the message if it starts with a "command" like
`.dosomething`.

* Move hasCommand method to MsgEvent

* Remove MsgEvent enum

* Update description