MsgHook v0.2.0 (#3)
* Make hooks only run when plugin is enabled

* Add property to check type of message event

Adds a property to check whether a message was newly sent or edited.
Also renames some interfaces.

* Add check to make sure __sentry_xhr__ exists

* Add message ID property to MsgHookEvent

Adds a message ID property that uses Promises to the MsgHookEvent
interface. This allows developers to track the ID's of newly sent as
well as edited messages.

* Add CHANGELOG.md

* Remove links from CHANGELOG.md

Removes links for versions from the CHANGELOG.md file. Links would've
been hard to properly maintain since branches are deleted after pull
requests are merged, and the merged commit hash can't be known in
advance to set the link before merging.